move dm related Activities to their own package
