+++ app/src/main/res/layout/activity_dms.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".activities.DirectMessages">
+    tools:context=".activities.directmessages.DirectMessages">
 
     <include
         android:id="@+id/toolbar"
